fix(agents): preserve lifecycle build boundaries
