+++ src/agents/subagents/registry/subagent-registry-lifecycle.ts
@@ -1,8 +1,6 @@
-import { sliceUtf16Safe } from "@openclaw/normalization-core/utf16-slice";
 import type { cleanupBrowserSessionsForLifecycleEnd } from "../../../browser-lifecycle-cleanup.js";
 import type { OpenClawConfig } from "../../../config/types.openclaw.js";
 import type { callGateway as defaultCallGateway } from "../../../gateway/call.js";
-import { formatErrorMessage, readErrorName } from "../../../infra/errors.js";
 import { runWithGatewayIndependentRootWorkAdmission } from "../../../process/gateway-work-admission.js";
 import type { DetachedTaskFindResult } from "../../../tasks/detached-task-runtime-contract.js";
 import type { AcceptedSessionSpawn } from "../../accepted-session-spawn.js";
@@ -82,24 +80,6 @@ export type SubagentLifecycleOptions = {
 export interface SubagentLifecycleCommonContext {
   readonly options: SubagentLifecycleOptions;
   newerGenerationOwnsSession(entry: SubagentRunRecord): boolean;
-}
-
-export function buildSafeLifecycleErrorMeta(error: unknown): Record<string, string> {
-  const message = formatErrorMessage(error);
-  const name = readErrorName(error);
-  return name ? { name, message } : { message };
-}
-
-export function maskLifecycleIdentifier(value: string, kind: "run" | "session"): string {
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return "unknown";
-  }
-  return kind === "session"
-    ? `${trimmed.split(":").slice(0, 2).join(":") || "session"}:…`
-    : trimmed.length <= 8
-      ? "***"
-      : `${sliceUtf16Safe(trimmed, 0, 4)}…${sliceUtf16Safe(trimmed, -4)}`;
 }
 
 export interface SubagentLifecycleCompletionContext extends SubagentLifecycleCommonContext {
